TileMaker

Function 3 of 3, the last step. This function grabs the Divs created by `DivMaker` and combines them into an html file.

TileMaker(MainTitle = "", Divs, FileName = "x", ShowDate = FALSE,
  localCSS = FALSE)

Arguments

MainTitle
Optional title for the whole set of titles

Divs
The Divs that you want inserted into the tile. If you have more than one div, use `paste(Div1,Div2)`

FileName
The filename the tile should spit out as, including the extension (which should always be html)

ShowDate
Optional boolean controlling whether the date should be included or suppressed

localCSS
Optional boolean to specify whether the bootstrap css file should be served from the internet, or if you have saved a local version. If you have saved a local version, make sure to download the "fonts" folder too, otherwise glyphicons won't work.

Value

Use this function to output an html file containing all the divs.If you would like just HTML code (suitable for inserting in a dashboard or another document, you can use the Divs.

References

Uses Twitter's awesome bootstrap V3

Examples

# ADD EXAMPLES HERE
Button1 <- ButtonMaker(Color = 2, Value = 3.57, Subtitle = "B")

#> Warning: These functions are provided for
#> compatibility with older versions of R only, and may be
#>               defunct as soon as of the next release.

Button2 <- ButtonMaker(Color = 3, Value = 13.7, Subtitle = "Nutritional value")

#> Warning: These functions are provided for
#> compatibility with older versions of R only, and may be
#>               defunct as soon as of the next release.

Div1 <- DivMaker(Title = "Quantativity factors", Buttons = paste(Button1, Button2))

#> Warning: These functions are provided for compatibility with older versions of R only,
#> and may be defunct as soon as of the next release.

Div2 <- DivMaker(Title = "Inverse proportions", Buttons = paste(Button2, Button1))

#> Warning: These functions are provided for compatibility with older versions of R only,
#> and may be defunct as soon as of the next release.

TileMaker(MainTitle = "Hello", Divs = paste(Div1, Div2), FileName = "example.html")

#> Warning: These functions are provided for compatibility with older versions of R only,
#> and may be defunct as soon as of the next release.

TileMaker(MainTitle = "Hello", Divs = paste(Div1, Div2))

#> Warning: These functions are provided for compatibility with older versions of R only,
#> and may be defunct as soon as of the next release.

#> [1] "<!DOCTYPE html><html lang=\"en\"><head>\n      <meta name=\"viewport\" content=\"width=device-width, initial-scale=1\"><link rel=\"stylesheet\" href=\n        \"http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/\n        bootstrap.min.css\"></head><body><h1>Hello</h1><div class=\"container\"><h2>Quantativity factors</h2><button type=\"button\" class=\"btn btn-warning btn-lg\"><h1>3.57</h1>B</button> <button type=\"button\" class=\"btn btn-danger btn-lg\"><h1>13.7</h1>Nutritional value</button></div> <div class=\"container\"><h2>Inverse proportions</h2><button type=\"button\" class=\"btn btn-danger btn-lg\"><h1>13.7</h1>Nutritional value</button> <button type=\"button\" class=\"btn btn-warning btn-lg\"><h1>3.57</h1>B</button></div></body></html>"

browseURL("example.html")