Div maker

This function takes buttons made by any of the solo or multi buttons and makes an a row (HTML `div`) suitable for inclusion in other HTML code, or for inclusion within the function of this package `finisher`.

div_maker(subtitle, textModifier, ...)

Arguments

subtitle

The text heading of this row of buttons

textModifier

Optional css category of "large" text. In this case, subtitle. Use css flags like "h2", "h3","p", etc. Default = "h1"

...

buttons to insert into the div elements.

Examples

div_maker(
  subtitle = "Quantativity factors", textModifier = "h1",
  solo_gradient_box(value = 70),
  solo_box(value = 34)
)
#> -- using target value of 100 --
#> <div class="container">
#>   <h1>Quantativity factors</h1>
#>   <a>
#>     <button class="btn btn-md btn-warning" color="warning" role="button">
#>       <h1>70</h1>
#>       
#>     </button>
#>   </a>
#>   <a>
#>     <button class="btn btn-md btn-info" color="info" role="button">
#>       <h1>34</h1>
#>       
#>     </button>
#>   </a>
#> </div>