multi_box

Create a tile that contains more than one value, icon and text

multi_box(
  icons = NULL,
  txt = NULL,
  values = NULL,
  title = NULL,
  size = "md",
  color = "info",
  link = NULL,
  number_zoom = 150,
  hover = NULL,
  ...
)

Arguments

icons

vector of Icons to display, Default: NULL

txt

Optional subtext that should appear under the value

values

vector of values to display, Default: NULL

title

Top title, Default: NULL

size

Optional size specified in the bootstrap css classes: "xs","sm","md","lg")

color

Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/ Choose from: "muted", "primary", "success", "info", "warning", "danger", Default: 'info'

link

Optional hyperlink to redirect to after a user click, Default: NULL

number_zoom

Optional magnification % for number vs normal text, Default: 150

hover

Optional tooltip, or text that will show up when a user rests their mouse over the tile, Default: NULL

...

add any other html code here

Value

an HTML object

Details

Allows for each button to contain several icon-number-text descriptions.

Examples

library(dplyr)
#> Warning: package 'dplyr' was built under R version 4.5.1
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
multi_box(
  values = c(21, 45), title = "Important <br>button",
  number_zoom = 300, icons = c("apple", "calendar"), color = "warning",
  txt = c("times", "reports")
) %>%
  finisher(divs = .)

  
    

    
      
    
  
if (FALSE) { # \dontrun{
if (interactive()) {
  # EXAMPLE1
}
} # }