solo_box

This function crafts the actual tile per se, including the specific aesthetic traits for each tile. This is the simple version where you explicitly state the color.

solo_box(value = NULL, txt = NULL, former = NULL, size = "md",
  icon = NULL, type = "info", link = NULL, units = NULL,
  hover = NULL, textModifier = "h1", ...)

Arguments

value
The numeric value you want to highlight (the main enchilada)

txt
Optional subtext that should appear under the value

former
The numeric old value to use for comparison to 'value'

size
Optional size specified in the bootstrap css classes: "xs","sm","md","lg")

icon
Optional glyphicon that should be displayed from http://getbootstrap.com/components/ you need only supply the name of thing you want, like "check"... not the full "gyphicon-check"

type
Optional bootstrap css element that governs the color. https://v4-alpha.getbootstrap.com/utilities/colors/ Choose from: "Muted", "Primary", "Success", "Info", "Warning", "Danger"

link
Optional hyperlink that should be followed on click

units
Optional units that should be displayed after Value

hover
Optional tooltip, or text that will show up when a user rests their mouse over the tile.

textModifier
Optional css category of "large" text. In this case, the icon, value and unit. In this case, title. Default=h1

...
Optional additional html elements

Examples

b1 <- solo_box(type = "warning", value = 3.57, txt = "B")
b2 <- solo_box(type = "danger", value = 13.7, txt = "Nutritional value")
b3 <- solo_box(type = "success", value = 1, txt = "Yumminess factor")
b4 <- solo_box(value = 3.57, former = 3, txt = "Times apple eaten", icon = "apple")
finisher(title = "straight buttons", divs = b1)
finisher(
  title = "with divs",
  divs = div_maker(
    subtitle = "boom",
    textModifier = "h1",
    div_maker(subtitle = "Boom", textModifier = "hi", b1, b2, b3)
  )
)

## Or taking advantage of the ability to change the textModifier:
finisher(
  title = "h4 modifier",
  divs = solo_box(
    value = 3, txt = "uh huh",
    former = 2, textModifier = "h4"
  )
)